Added BCrypt for user passwords
Now, the passwords are not stored as plaintext.  Instead, they use the
recommended BCrypt encryption.  I had to extend the length of the
password field in the database.

The standalone.xml still thinks there is plaintext passwords; but this
isn’t a big deal if we just ignore the authentication for calling EJB
methods…not a terribly risky option.