More work related to joining and quitting games
Added System logging for the EJB.
Added a couple sample users, just in case I need to repopulate the DB
from scratch.
Fixed auto generation of IDs for the following tables: GameState,
Player, User
* Please note: this change would be overwritten if the JPA classes are
auto generated!
Added Ability to quit a game
Added some logic to the “Game_Play.jsp” view, which is where the game
state should be rendered.  This can be the destination view for many
different URL requests.
Cleaned the menu to better reflect whether a player is currently in an
active game.